+++ api-builder-plugin-fn-aws-lambda/src/actions.js
@@ -1,5 +1,3 @@
-const aws = require("aws-sdk");
-
 /**
  * Action method.
  * @param {object} req - The flow request context passed in at runtime.  The
@@ -22,35 +20,17 @@ function invokeLambda(req, outputs, options) {
   const payload = req.params.payload;
   const asynchronous = req.params.asynchronous;
 
-  const awsConfig = this.pluginConfig.aws;
-
   var logType = 'None';
+  const lambda = this.lambdaClient;
 
   if (!func) {
 		options.logger.error('The func parameter is missing.');
 		return outputs.error(null, {message: 'Missing required parameter: func'});
 	}
 
-  if(typeof awsConfig.credentials.region === 'undefined' || awsConfig.credentials.region == 'PROVIDE_YOUR_AWS_REGION' ||
-    typeof awsConfig.credentials.accessKeyId === 'undefined' || awsConfig.credentials.accessKeyId == 'PROVIDE_YOUR_AWS_ACCESS_KEY_ID' ||
-    typeof awsConfig.credentials.secretAccessKey === 'undefined' || awsConfig.credentials.secretAccessKey == 'PROVIDE_YOUR_AWS_SECRET')
-  {
-    options.logger.error('Your AWS configuration is incomplete. Please check conf/aws-lambda.default.js');
-  	return outputs.error(null, {message: 'Your AWS configuration is incomplete. Please check conf/aws-lambda.default.js'});
-  }
-
   if(logResult) {
     logType = 'Tail';
   }
-
-  awsCredentials = {
-    region: awsConfig.credentials.region,
-    accessKeyId: awsConfig.credentials.accessKeyId,
-    secretAccessKey: awsConfig.credentials.secretAccessKey
-  };
-  aws.config.update(awsCredentials);
-
-  var lambda = new aws.Lambda({region: awsConfig.credentials.region, apiVersion: '2015-03-31'});
 
   var lambdaPayload = payload;
   if(typeof payload === 'object') {
@@ -78,24 +58,18 @@ function invokeLambda(req, outputs, options) {
 
 function callLambda(options, outputs, lambda, pullParams, logResult, asynchronous) {
 
-  lambdaPromise = lambda.invoke(pullParams).promise();
-
-  lambdaPromise.then(
-    function(data) {
-      if(logResult) {
-        options.logger.debug(Buffer.from(data.LogResult, 'base64').toString('utf8'));
-      }
+  lambdaPromise = lambda.invoke(pullParams, function(err, data) {
+    if(err) {
+      return outputs.error(null, {message: error});
+    } else {
       if(asynchronous && data.StatusCode == 202) {
         return outputs.next(null, 'Accepted');
       } else {
+        debugger;
         return outputs.next(null, JSON.parse(data.Payload));
-      }
-    },
-    function(error) {
-      options.logger.error(error);
-      return outputs.error(null, error);
+      }      
     }
-  )
+  });
 }
 
 module.exports = {